fixed64: hand-fix batch 1 -- angle->fixed, double<->fixed, misc
35 per-site fixes with intent:
  - angle_t->fixed_t stored/compared: wrap in fixed_t(...) (zero-extend,
    matches old int64=uint32) -- AngleDelta, pitch +=/-=, shootangle, etc.
  - double<->fixed geometry (nodebuild/p_sectors/p_slopes ClosestPoint,
    r_utility finetangent): double() on the fixed operand.
  - int accumulators += Fixed -> (int); gl_texture Scale uses zx::raw of
    the Fixed factor (int64 overload); gl_sprite float(center_y); DScroller
    dy fixed_t(0) to disambiguate the ctor.
  - Reverted a codemod over-wrap in p_acs (20 nested (int) -> one).
Behavior-preserving; flag-off green; tests green.

+++ src/zandronum/src/g_shared/sbarinfo_commands.cpp
@@ -1431,10 +1431,10 @@ class CommandDrawNumber : public CommandDrawString
 					//Hexen counts basic armor also so we should too.
 					if(statusBar->armor != NULL)
 					{
-						num += FixedMul(statusBar->armor->SavePercent, 100*FRACUNIT);
+						num += (int)FixedMul(statusBar->armor->SavePercent, 100*FRACUNIT);
 					}
 					if(value == ARMORCLASS)
-						num /= (5*FRACUNIT);
+						num /= (int)(5*FRACUNIT);
 					else
 						num >>= FRACBITS;
 					break;
